Move CnAlgo
