+++ src/crypto/cn/CnAlgo.h
