+++ src/crypto/cn/CryptoNight_arm.h
@@ -28,8 +28,8 @@
 #define XMRIG_CRYPTONIGHT_ARM_H
 
 
-#include "base/crypto/CnAlgo.h"
 #include "base/crypto/keccak.h"
+#include "crypto/cn/CnAlgo.h"
 #include "crypto/cn/CryptoNight_monero.h"
 #include "crypto/cn/CryptoNight.h"
 #include "crypto/cn/soft_aes.h"
